feat: add remove contract action
